+++ plugins/@grouparoo/mailchimp/src/initializers/plugin.ts
@@ -5,7 +5,7 @@ import { test } from "./../lib/test";
 import { parallelism } from "./../lib/parallelism";
 
 import emailDestination from "../lib/export/connection";
-import extidDestination from "../lib/export-external-id/connection";
+import extidDestination from "../lib/export-id/connection";
 import importSource from "../lib/import/connection";
 
 const packageJSON = require("./../../package.json");
